Fix ROM_0 being full

+++ buildrom.bat
@@ -1,6 +1,9 @@
 @echo off
-rmdir /s /q temp
-rmdir /s /q bin
-start make clear && make
-gbdk\bin\romusage.exe bin/MiJuego.gb
+if exist temp rmdir /s /q temp
+if exist bin rmdir /s /q bin
+call make clean
+call make
+if exist bin\MiJuego.gb (
+    gbdk\bin\romusage.exe bin/MiJuego.gb
+)
 pause
